Minor bugfix
User can set their end date to be before the start date when creating
"Event" tasks.

This obviously defies logic and should be changed.

Creating a new exception to be thrown, and throwing it when the end date
is before the start date, fixes the problem.

This ensure that the exception can be caught as per normal, and the
message should inform the user that they might need to get their eyes
(or brain) checked.