Improve help functionality
Help functionality is purely text-based and unintuitive, with no
examples shown to the user.

This makes the help not very helpful, and users may have to keep
scrolling up and down to copy and find the information they need.

Let's make the app open a new window containing the user guide when
the user asks for help.

As we already have a polished user guide (README.md), it would be
easier and better to just reuse that for the internal help function,
as updating the guide as more functionalities are added is much easier
this way. In addition, the user may be using the app offline, so just
opening the app's website is not good.

A new dependency, FlexMark, was added to convert .md to HTML (as JavaFX
does not natively support reading .md files), and the HTML file is
opened in a new WebView using JavaFX.

In addition, updated the parser to read user inputs for single-word
commands (help, list etc) even if there are leading or trailing
whitespaces, in the spirit of fat fingers.

Let's,
* Make sure the new dependency is properly bundled into the .jar file
  when running shadowJar
* Test that the user guide opens correctly in other systems