Redesign user guide
User guide was relying on too many different dependencies, inflating
the file size of the fat .jar file.

This is a waste of space, as a ~10mb file is inflated to ~110mb just to
accommodate the additional dependencies. Repeatedly converting a single
.md file to HTML is inefficient and unnecessary too.

Creating a whole native UI in JavaFX for the user guide solves the
dependency issue.

Doing it using .fxml and .css files (like the dialog box) removes
the need for WebView and Media dependencies in javaFX, which no longer
come bundled with Java by default. This saves a lot of storage space,
cutting it down to ~10%.